refactor: extract component logic into custom hooks and update components to utilize them.
