+++ src/logic/useContactForm.js
@@ -1,0 +1,26 @@
+import { useState } from "react";
+import { useToast } from "@/hooks/use-toast";
+
+export const useContactForm = () => {
+  const { toast } = useToast();
+  const [isSubmitting, setIsSubmitting] = useState(false);
+
+  const handleSubmit = (e) => {
+    e.preventDefault();
+
+    setIsSubmitting(true);
+
+    setTimeout(() => {
+      toast({
+        title: "Message sent!",
+        description: "Thank you for your message. I'll get back to you soon.",
+      });
+      setIsSubmitting(false);
+    }, 1500);
+  };
+
+  return {
+    isSubmitting,
+    handleSubmit,
+  };
+};
